requirements.txt
requests>=2.25.0
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
flask>=2.0.0
django>=4.0.0
fastapi>=0.68.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
redis>=4.0.0
celery>=5.2.0
pydantic>=1.8.0
click>=8.0.0
rich>=12.0.0
typer>=0.4.0
httpx>=0.23.0
aiohttp>=3.8.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
isort>=5.10.0
beautifulsoup4>=4.10.0
lxml>=4.6.0
openpyxl>=3.0.0
pillow>=8.3.0
boto3>=1.18.0
google-cloud-storage>=2.0.0
azure-storage-blob>=12.8.0
paramiko>=2.7.0
cryptography>=3.4.0
jwt>=1.3.0
passlib>=1.7.0
bcrypt>=3.2.0